requirements.txt
newspaper3k
Requests
flask
newspaper3k
python-dotenv
gunicorn